Change: MenuBar.vueを消す
